backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.1.0
openai>=1.12.0
python-dotenv>=1.0.0
jinja2>=3.1.3
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.26.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
email-validator>=2.2.0
authlib>=1.3.0